Test the GitHub CI workflow on Ubuntu 24.04
Our GitHub CI workflow run on ubuntu-latest, which is an alias for the
ubuntu-22.04 runner image. Starting on 2024-12-05, ubuntu-latest will be
progressively migrated to an Ubuntu 24.04 image, which has a different
set of installed packages than the current image.[1]

Make sure our workflow is compatible with the new image by testing it
against an explicitly requested ubuntu-24.04 image.

[1] https://github.blog/changelog/2024-11-05-notice-of-breaking-changes-for-github-actions/